Move aria helpers from components level to utils
